Refactoring PlaceCard
