+++ lib/main.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:places/common/error_handler.dart';
 import 'package:places/data/api/api_client.dart';
+import 'package:places/data/blocs/filtered_places/bloc/filtered_places_bloc.dart';
 import 'package:places/data/interactor/place_interactor.dart';
 import 'package:places/data/interactor/search_interactor.dart';
 import 'package:places/data/repository/place_repository.dart';
@@ -10,26 +12,19 @@ import 'package:places/ui/screens/res/themes.dart';
 import 'package:places/ui/screens/splash_screen.dart';
 import 'package:provider/provider.dart';
 
-final api = ApiClient();
-
 void main() {
   runApp(App());
 }
 
 class App extends StatelessWidget {
+  final api = ApiClient();
+
   @override
   Widget build(BuildContext context) {
     return MultiProvider(
       providers: [
-        Provider<SearchInteractor>(
-          create: (_) => SearchInteractor(SearchRepository(api)),
-        ),
-        Provider<PlaceInteractor>(
-          create: (_) => PlaceInteractor(PlaceRepository(api)),
-        ),
-        Provider<StandardErrorHandler>(
-          create: (_) => StandardErrorHandler(),
-        )
+        BlocProvider<FilteredPlacesBloc>(
+            create: (context) => FilteredPlacesBloc(SearchRepository(api))),
       ],
       child: MaterialApp(
         title: 'Places',
