+++ lib/ui/screens/main_screen.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:places/ui/screens/settings_screen.dart';
-import 'package:places/ui/screens/place_list_screen/place_list_screen.dart';
-import 'package:places/ui/screens/place_map_screen.dart';
-import 'package:places/ui/screens/favorites_screen.dart';
+// import 'package:places/ui/screens/settings_screen.dart';
+import 'package:places/ui/screens/place_list_screen.dart';
+// import 'package:places/ui/screens/place_map_screen.dart';
+// import 'package:places/ui/screens/favorites_screen.dart';
 import 'package:places/ui/widgets/place_bottom_nav_bar.dart';
 
 class MainScreen extends StatefulWidget {
@@ -16,7 +16,6 @@ class MainScreen extends StatefulWidget {
 class _MainScreenState extends State<MainScreen> with TickerProviderStateMixin {
   int initialIndex = 0;
   late TabController _tabController;
-  
 
   @override
   void initState() {
@@ -47,27 +46,22 @@ class _MainScreenState extends State<MainScreen> with TickerProviderStateMixin {
 
   @override
   Widget build(BuildContext context) {
-    final isPortrait =
-        MediaQuery.of(context).orientation == Orientation.portrait;
-
     return Scaffold(
       body: TabBarView(
         controller: _tabController,
-        children: const [
+        children: [
           PlaceListScreen(),
           // PlaceMapScreen(),
           // FavoritesScreen(),
           // SettingsScreen(),
         ],
       ),
-      bottomNavigationBar: isPortrait
-          ? PlaceBottomNavBar(
-              currentIndex: initialIndex,
-              onSelectTab: (index) {
-                onSelectTab(index);
-              },
-            )
-          : const SizedBox.shrink(),
+      bottomNavigationBar: PlaceBottomNavBar(
+        currentIndex: initialIndex,
+        onSelectTab: (index) {
+          onSelectTab(index);
+        },
+      ),
     );
   }
